+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 First Release.
 
+## [0.0.4]
+
+-   windows open cmd/wsl/bash termail bug fix
+-   onSaveEvents options: format apex, format trigger ...
+
 ## [0.0.3]
 
 -   Support using WSL/git bash/Msys2/MingW64/MingW32 to develop sfdx.